The most significant changes involve the refactoring of the `Cannonball` and `Ship` classes, the introduction of the `InputHandler` class, and the modification of the `Hashmap` class.
In the `Cannonball` class, the constructor was simplified and the texture and scale setting for the sprite was moved to the `shootCannonballs` method. The `texture`, `cannonballScale`, and `textures` members were also removed from this class.

The `Ship` class underwent several changes including the renaming of the `shipConfig` member to `ShipConfig`, and the removal of the `window`, `collisionRect`, and `rotation` members. The rotation is now retrieved directly from the sprite in the `setHealthBarPosition` method.

The `InputHandler` class was introduced and added as a member to the `GlobalValues` class. This new class is used for handling mouse and keyboard input in the `updateAndDraw` method in `Ship_PG.cpp` and is updated at the start of each frame in the `createWorld` method in `World_PG.cpp`.

The `Hashmap` class was modified to use a custom hash function for `std::pair<int, int>` and the `generateKey` method and `hashmap` member were changed to use `std::pair<int, int>` instead of `std::string`.

Other changes include the renaming of the `collisionSound` member to `sound` in the `SoundManager_PG` class, the reduction of the `cooldown` member from `2.5f` to `0.1f` in the `shipCannonHandler_PG` class, and the addition of the `cannonballScale` and `textures` members to the `ShipCannonHandler` class.

Finally, the `SFML Project 1.vcxproj` and `SFML Project 1.vcxproj.filters` files were updated to include the new `InputHandler_PG.h` and `InputHandler_PG.cpp` files.