The most significant changes involve the introduction of a new class `GlobalHashmapHandler` which is a singleton that contains all the hashmaps for the game. This class includes methods to get instances of `LandMassHashmap` and `ShipHashmap`. This change has led to updates in several files to include this new class and remove the individual hashmap files. The `CollisionManager_PG.cpp`, `EnemyShipHandler_PG.cpp`, `LandMassHandler_PG.cpp`, and `ShipHashmap_PG.cpp` files have been updated to use the global hashmaps from the `GlobalHashmapHandler` instance instead of the local hashmaps.
The `Map::getRandomPositions` method in `Map.cpp` has been updated to include a condition that checks if the size of `samplePoints` is less than `numPoints` in the while loop. This ensures that the loop does not run indefinitely if the number of points generated is less than the desired number.

The `EnemyShipHandler_PG.h` file has been updated to include the `GlobalHashmapHandler.h` file and remove the `ShipHashmap_PG.h` file. The destructor of the `EnemyShipHandler` class has been updated to remove all enemy ships and remove them from the hashmap.

The `ShipHashmap_PG.h` file has been updated to include a `maxDistance` parameter in the `findEnemyShipsNearShip` method.

The `EnemyShipInputHandler_PG.cpp` file has been updated to only fire if the player is within a certain firing range.

The `Ship_PG.cpp` file has been updated to generate a random number between 1 and 5 if the ship level is random.

The `World_PG.cpp` file has been updated to set up the world with a certain number of enemy ships based on the map size.

The `SFML Project 1.vcxproj` and `SFML Project 1.vcxproj.filters` files have been updated to include the `GlobalHashmapHandler.h` file.

Finally, the `GlobalHashmapHandler.h` file has been added which includes the definition of the `GlobalHashmapHandler` class.