The most significant changes involve the targeting system in the `ShipGroup_PG.h` and `ShipGroup_PG.cpp` files. The `setTarget` method has been replaced with `addTarget`, which takes an `EnemyShip` object as a parameter, allowing the `ShipGroup` class to target specific enemy ships. The `target` variable has also been replaced with `targetShip`, a pointer to an `EnemyShip` object. In the `updateGroup` method, the target position is now set to the position of the target ship.
In the `EnemyShipMovementHandler_PG.h` file, the `EnemyShip_PG.h` file is now included, allowing the `EnemyShipMovementHandler` class to access the `EnemyShip` class.

In the `EnemyShipHandler_PG.cpp` file, the condition for entering combat has been changed to always initiate combat. The target for the ship group is now set to a random enemy ship, and ships that are no longer nearby are removed from combat.

In the `EnemyShipMovementHandler_PG.cpp` file, the calculation for `travelDirection` has been modified, and the conditions for changing `travelDirection` have been adjusted. These changes affect how the enemy ship moves towards its target. The `setSpriteRotation` method has been modified to adjust the turning speed of the sprite based on whether it is moving towards a target.

Three getter methods have been added to the `ShipGroup_PG.h` file: `getGroupIDsInteractedWith`, `getShipIDsCombatting`, and `calculateCohesion`. These methods return the IDs of groups that the ship has interacted with, the IDs of ships that the ship is currently combating, and the cohesion of the ship group, respectively.

Lastly, the debugging output in `ShipGroup_PG.cpp` has been modified to display the position of the target ship.